+++ static/app/views/starfish/landing/views/starfishView.tsx
@@ -1,5 +1,8 @@
+import styled from '@emotion/styled';
 import {Location} from 'history';
 
+import {PerformanceLayoutBodyRow} from 'sentry/components/performance/layouts';
+import {space} from 'sentry/styles/space';
 import {Organization, Project} from 'sentry/types';
 import EventView from 'sentry/utils/discover/eventView';
 import {usePageError} from 'sentry/utils/performance/contexts/pageError';
@@ -7,7 +10,7 @@ import {PerformanceDisplayProvider} from 'sentry/utils/performance/contexts/perf
 
 import Table from '../../table';
 import {PROJECT_PERFORMANCE_TYPE} from '../../utils';
-import {ChartRow} from '../widgets/components/widgetChartRow';
+import WidgetContainer from '../widgets/components/widgetContainer';
 import {PerformanceWidgetSetting} from '../widgets/widgetDefinitions';
 
 type BasePerformanceViewProps = {
@@ -22,14 +25,28 @@ export function StarfishView(props: BasePerformanceViewProps) {
   return (
     <PerformanceDisplayProvider value={{performanceType: PROJECT_PERFORMANCE_TYPE.ANY}}>
       <div data-test-id="starfish-view">
-        <ChartRow
-          chartCount={1}
-          chartHeight={180}
-          {...props}
-          allowedCharts={[PerformanceWidgetSetting.DB_HTTP_BREAKDOWN]}
-        />
+        <StyledRow minSize={200}>
+          <WidgetContainer
+            allowedCharts={[PerformanceWidgetSetting.DB_HTTP_BREAKDOWN]}
+            chartCount={1}
+            chartHeight={180}
+            eventView={props.eventView}
+            location={props.location}
+            key={0}
+            withStaticFilters
+            index={0}
+            defaultChartSetting={PerformanceWidgetSetting.DB_HTTP_BREAKDOWN}
+            rowChartSettings={[PerformanceWidgetSetting.DB_HTTP_BREAKDOWN]}
+            setRowChartSettings={() => {}}
+          />
+        </StyledRow>
+
         <Table {...props} setError={usePageError().setPageError} />
       </div>
     </PerformanceDisplayProvider>
   );
 }
+
+const StyledRow = styled(PerformanceLayoutBodyRow)`
+  margin-bottom: ${space(2)};
+`;
